@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+7
 
 environment:
-  sdk: ">=2.11.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   auto_size_text: ^2.1.0
@@ -27,7 +27,7 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_markdown: ^0.6.4
-  flutter_riverpod: ^0.14.0+3
+  flutter_riverpod: ^1.0.3
   flutter_svg: ^0.22.0
   fluttertoast: ^8.0.8
   firebase_core: ^1.6.0
@@ -36,14 +36,13 @@ dependencies:
   hive: ^2.0.4
   hive_flutter: ^1.1.0
   image_gallery_saver: ^1.6.9
-  lint: ^1.5.3
   open_file: null
   path_provider: ^2.0.2
   permission_handler: ^8.1.4+2
   receive_sharing_intent: ^1.4.5
   share: ^2.0.4
   url_launcher: ^6.0.9
-  video_player: ^2.1.12
+  video_player: ^2.2.19
 
 dev_dependencies:
   build_runner: ^2.1.0
@@ -70,12 +69,21 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
-    - assets/bitmap2.svg
-    - assets/empty.svg
-    - assets/logo2.png
-    - assets/1.png
-    - assets/2.png
-    - assets/3.png
+
+  fonts:
+  - family: Quicksand
+    fonts:
+      - asset: assets/fonts/Quicksand-Medium.ttf
+        weight: 500
+      - asset: assets/fonts/Quicksand-SemiBold.ttf
+        weight: 600
+      - asset: assets/fonts/Quicksand-Bold.ttf
+        weight: 700
+
+  - family: Cookie
+    fonts:
+      - asset: assets/fonts/Cookie-Regular.ttf
+        weight: 400
 
 flutter_native_splash:
   background_image: assets/logo2.png
@@ -89,17 +97,12 @@ flutter_native_splash:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
+
+    # - family: Trajan Pro
+    #   fonts:
+    #     - asset: fonts/TrajanPro.ttf
+    #     - asset: fonts/TrajanPro_Bold.ttf
+    #       weight: 700
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
