@@ -1,15 +1,17 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:permission_handler/permission_handler.dart';
-import 'package:reels_downloader/model/ads/ad_model.dart';
-import 'package:reels_downloader/model/photo/photo_model.dart';
-import 'package:reels_downloader/model/useraccounts/user_model.dart';
-import 'package:reels_downloader/model/video/video_model.dart';
-import 'view/mainpage/page/mainpage.dart';
+
+import 'model/ads/ad_model.dart';
+import 'model/photo/photo_model.dart';
+import 'model/useraccounts/user_model.dart';
+import 'model/video/video_model.dart';
+import 'view/screens/mainpage.dart';
 
 const String videoBox = 'videomodel';
 const String userBox = 'usermodel';
@@ -31,7 +33,7 @@ Future<void> main() async {
   await Hive.openBox<UserModel>(userBox);
   await Hive.openBox<VideoModel>(videoBox);
   await Hive.openBox<PhotoModel>(photoBox);
-  await AdServices.createRewardedAd();
+  // await AdServices.createRewardedAd();
   runApp(ProviderScope(child: LandingPage()));
 }
 
@@ -50,12 +52,33 @@ class _LandingPageState extends State<LandingPage> {
 
   @override
   Widget build(BuildContext context) {
+    SystemChrome.setSystemUIOverlayStyle(
+        const SystemUiOverlayStyle(statusBarColor: Colors.transparent));
     getPermission();
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: SafeArea(
-        child: MainPage(),
+      theme: ThemeData(
+        fontFamily: 'Quicksand',
+        textTheme: TextTheme(
+          headlineMedium: TextStyle(
+            fontFamily: 'Cookie',
+            fontSize: 64,
+            color: Color.fromARGB(255, 240, 111, 111),
+            fontWeight: FontWeight.w700,
+          ),
+          bodySmall: TextStyle(
+            fontSize: 14,
+            color: Colors.white,
+            fontWeight: FontWeight.w600,
+          ),
+          bodyMedium: TextStyle(
+            fontSize: 16,
+            color: Colors.white,
+            fontWeight: FontWeight.w700,
+          ),
+        ),
       ),
+      home: MainPage(),
     );
   }
 }
